Reorder and rename some libp2p internals
Reorder and rename some libp2p internals to have a semantic order:
- Move everything related to the message dispatcher to a submodule.
  - This include the move of the `message_codec` to the newly crated
    `dispatcher` submodule.
- Move the `codecs` submodule inside the `discovery` submodule since
  it is only used there.

Also changed all the `use` statements to match the new hierarchy.
