+++ network-libp2p/src/discovery/message_codec/header.rs
