+++ network-libp2p/src/discovery/mod.rs
@@ -1,4 +1,5 @@
 pub mod behaviour;
 pub mod handler;
+pub mod message_codec;
 pub mod peer_contacts;
 pub mod protocol;
