+++ network-libp2p/src/dispatch/message_dispatch.rs
@@ -15,7 +15,7 @@ use tokio_util::codec::Framed;
 
 use beserial::{Deserialize, Serialize};
 
-use crate::codecs::{
+use super::codecs::{
     tokio_adapter::TokioAdapter,
     typed::{Error, Message, MessageCodec, MessageType},
 };
